ADEN-12446 Move N&R platforms to news-and-ratings dir
