+++ src/platforms/news-and-ratings/comicvine/setup/dynamic-slots/comicvine-dynamic-slots.setup.ts
