+++ src/platforms/news-and-ratings/gamefaqs/ad-context.ts
