+++ src/platforms/news-and-ratings/gamefaqs/bidders/appnexus.ts
