+++ src/platforms/news-and-ratings/gamefaqs/bidders/index-exchange.ts
