+++ src/platforms/news-and-ratings/gamefaqs/bidders/medianet.ts
