+++ src/platforms/news-and-ratings/gamefaqs/bidders/wikia.ts
