+++ src/platforms/news-and-ratings/gamefaqs/gamefaqs-platform.ts
@@ -1,20 +1,12 @@
 import { Injectable } from '@wikia/dependency-injection';
-import {
-	bidders,
-	communicationService,
-	context,
-	eventsRepository,
-	utils,
-	ProcessPipeline,
-} from '@wikia/ad-engine';
+import { bidders, communicationService, eventsRepository, ProcessPipeline } from '@wikia/ad-engine';
 import { bootstrapAndGetConsent, gptSetup } from '@platforms/shared';
 
-import { basicContext } from './ad-context';
 import { GamefaqsSlotsContextSetup } from './setup/context/slots/gamefaqs-slots-context.setup';
 import { GamefaqsDynamicSlotsSetup } from './setup/dynamic-slots/gamefaqs-dynamic-slots.setup';
-import { NewsAndRatingsTargetingSetup } from '../shared-news-and-ratings/context/targeting/news-and-ratings-targeting.setup';
+import { NewsAndRatingsTargetingSetup } from '../shared';
 import { GamefaqsPrebidConfigSetup } from './setup/context/prebid/gamefaqs-prebid-config.setup';
-import { GamefaqsTargetingSetup } from './setup/context/targeting/gamefaqs-targeting.setup';
+import { NewsAndRatingsBaseContextSetup } from '../shared';
 
 @Injectable()
 export class GamefaqsPlatform {
@@ -22,13 +14,9 @@ export class GamefaqsPlatform {
 
 	execute(): void {
 		this.pipeline.add(
-			() => context.extend(basicContext),
-			() => context.set('custom.dfpId', this.shouldSwitchGamToRV() ? 22309610186 : 5441),
-			() => context.set('src', this.shouldSwitchSrcToTest() ? ['test'] : context.get('src')),
-			// TODO: to decide if we want to call instant-config service for the first releases?
+			NewsAndRatingsBaseContextSetup,
 			() => bootstrapAndGetConsent(),
 			NewsAndRatingsTargetingSetup,
-			GamefaqsTargetingSetup,
 			GamefaqsSlotsContextSetup,
 			GamefaqsDynamicSlotsSetup,
 			// TODO: add targeting setup once we have idea of page-level and slot-level targeting
@@ -42,13 +30,5 @@ export class GamefaqsPlatform {
 		);
 
 		this.pipeline.execute();
-	}
-
-	private shouldSwitchGamToRV() {
-		return utils.queryString.get('switch_to_rv_gam') === '1';
-	}
-
-	private shouldSwitchSrcToTest() {
-		return utils.queryString.get('switch_src_to_test') === '1';
 	}
 }
