+++ src/platforms/news-and-ratings/gamefaqs/index.ts
