+++ src/platforms/news-and-ratings/gamefaqs/setup/dynamic-slots/gamefaqs-dynamic-slots.setup.ts
