+++ src/platforms/news-and-ratings/gamespot/ad-context.ts
