+++ src/platforms/news-and-ratings/gamespot/index.ts
