+++ src/platforms/news-and-ratings/gamespot/setup/context/slots/gamespot-slots-context.setup.ts
