+++ src/platforms/news-and-ratings/gamespot/setup/dynamic-slots/gamespot-dynamic-slots.setup.ts
