+++ src/platforms/news-and-ratings/giantbomb/ad-context.ts
