+++ src/platforms/news-and-ratings/giantbomb/index.ts
