+++ src/platforms/news-and-ratings/giantbomb/setup/context/slots/giantbomb-slots-context.setup.ts
