+++ src/platforms/news-and-ratings/giantbomb/setup/dynamic-slots/giantbomb-dynamic-slots.setup.ts
