+++ src/platforms/news-and-ratings/metacritic/ad-context.ts
