+++ src/platforms/news-and-ratings/metacritic/index.ts
