+++ src/platforms/news-and-ratings/metacritic/setup/dynamic-slots/metacritic-dynamic-slots.setup.ts
