+++ src/platforms/news-and-ratings/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './context/base/news-and-ratings-base-context.setup';
+export * from './context/targeting/news-and-ratings-targeting.setup';
