+++ src/platforms/news-and-ratings/shared/tsconfig.json
@@ -1,0 +1,9 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"paths": {
+			"@wikia/ad-engine": ["src/index.ts"]
+		}
+	},
+	"include": ["."]
+}
